Fix invalid cache state
Cache ended up in invalid state when there was categories used in the
installer which were not all defaultly selected. Cache
was only party updated when using CLI command search, and
the cache was populated only with Updates.xml and repository.txt. In the
next installer or MT launch, installer thought the cache was up to date
as the Updates.xml was found and did not know that the metadata was not
fetched. This lead to different error messages, like 'Cannot open script
file' or 'Cannot open the requested license file' etc.
Fixed so that metadata is fetched also when using CLI command search

Task-number: QTIFW-2998
Change-Id: I1cdc446469f38cc9e6ddc71f512105fc5d10e74c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>