Verify meta file integrity from cache
Calculate and store a checksum for each included file when extracing
meta.7z archives. When the respective meta item will be loaded from
cache on the next run, each file referenced in its Updates.xml will
be checked from disk, that it exist and that its checksum matches
the one calculated when the file was originally extracted.

If the files in cache are not intact, the meta item will be refreshed
(downloaded and extracted again).

Also bump the IFW cache format version, as the existing caches need
to be repopulated to include the new checksum files.

Task-number: QTIFW-3023
Change-Id: I6a8431ffb52cfff69e3cb84cccd4b3d5a1fe21da
Reviewed-by: Katja Marttila <katja.marttila@qt.io>