Fix wrong extract content
If same component is available from offline payload and from
online, both offline and online payload were extracted if the online
version number was higher.
Fixed so that available archives are filtered based on version number.
We read the offline resource content in the beginning of install when
we don't know that online version provides the same component with
higher version number. Moving the offline resource reading to later
phase would be another option for the fix but feels much more risky
solution.

Task-number: QTIFW-3010
Change-Id: Id4c57a6e72e41745f4d81d707c4f32414a15005e
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>