Fix license undo operation
Licenses value write was accidently changed from lower case to upper
case, which caused that the licenses undo operation did not find a value
for the license undo. Fixed the value settings, and also removed the
error dialog as there is no point in showing that in undo operation as
we are removing the component.

Task-number: QTIFW-3013
Change-Id: I402658a05b904454da045ba3fe38c21cf4c3dd4b
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>