Coin: Switch to the new packaging tools scripts location
Scripts were moved from qtsdk.git to tqtc-qtsdk.git.
Adjust the coin instructions to use the new location.

This will also update the packaging scripts, so the new macOS
version with Xcode 13+ must be in use before it can be merged.

With the new scripts in use, the .dmg disk image will be
signed in addition to the .app bundle by default.

Also, the notarization tool is switched from soon to be
deprecated 'altool' to the new 'notarytool'.

Task-number: QTQAINFRA-5505
Change-Id: I63046761eb19e833eed414e19c17402a5694b556
Reviewed-by: Katja Marttila <katja.marttila@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>