Consider full length of 'key=value' string in user arguments
In 'key=value' pairs passed to IFW on command line, 'value' can be
any string, including base64-encoded string which may contain 1 or 2
trailing '=' chars

The current algorithm of splitting 'key=value' string into sections
using '=' as separator and assigning field[1,1] to 'value' works in
most common case (no further '=' chars in 'value'). It fails sometimes
if 'value' is base64-encoded string as it cuts off the portion of
the string at 2nd '=' char and beyond, thereby discards trailing '='s,
which are significant. This invalidates the string.

To avoid this, we use QString.section to return fields[1,-1]
(i.e. everything starting from field1 to the end of the string)
and assign the result to 'value'

Change-Id: Ide4f1510170ad4685b5bb40a96d88898476553ea
Reviewed-by: Katja Marttila <katja.marttila@qt.io>