Add support for component aliases
Introduce concept of component aliases, which act as an alternative way
for referring a set of related components.

Component aliases are declared in an alias definition file, which is
included to the created installer's binary layout as a resource. The
file lists the available aliases, including metadata - such as name,
version, and description - and the list of components and other aliases
the alias requires.

Aliases can be referred only from the CLI for the time being, with the
supported commands 'install' and 'search'.

Task-number: QTIFW-2978
Change-Id: I281f171cc7d932ce496051d7090ae169a4709eec
Reviewed-by: Katja Marttila <katja.marttila@qt.io>