Prevent uninstalling components when error occurs
When error occurs during component fetch, for example cache is in use or
another metadata parsing error, installer allowed to continue
to uninstall existing components. This has confused users, as there is no
place to show error message and users had no idea why not all components
were visible in the tree. The error message was only written to logs.

Fixed this by showing the error message in 'Add or remove/Update' -page
and not allowing to continue to uninstall existing components.

Task-number: QTIFW-3069
Change-Id: I9f7d96d04b150c319b01d6e61430805128d3c31a
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>