DownloadArchivesJob: allow download retries for command line runs
Previously, on network errors or hash verification failures, the
download job would be canceled when running the installer from CLI. This
was to prevent infinite retries in case the automatic or default answer
for the error message boxes was to retry.

Add a retry counter and allow retrying the failed downloads from
CLI runs until the max count is reached.

Task-number: QTIFW-3098
Change-Id: Iee95f0eb4e47af89c74ffdd9ad6026bda6c87664
Reviewed-by: Katja Marttila <katja.marttila@qt.io>