Allow setting temporary repositories for offline-only installers
As the settings button can be configured to be visible for offline-only
installers, the user may configure additional repositories in addition
to the embedded binary content. Those repositories were not considered
for the metadata fetch, however, so components from those repositories
could not be installed.

Allow fetching metadata from temporary or user configured repositories regardless of if the installer is created as "offline-only". Also fix
required temporary space calculation when content is installed
simultaneously from both embedded to the installer binary and remote
repositories.

Task-number: QTIFW-3078
Change-Id: I4effa365e856ca009588ac8768f8b6c30b4b0373
Reviewed-by: Katja Marttila <katja.marttila@qt.io>