Disable http2 protocol
By default, installer build with Qt6 enables http2 protocol. Disabling
http2 protocol for now so that the behavior is the same when building
with Qt5. Currently some cdn's response with "GOAWAY received, cannot
start a request" -messages, the problem needs to be solved first before
http2 protocol can be enabled.

Change-Id: I0cad5fef4fe299d21ee66489583cd4411f6bd7c8
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>