Fix allowing selecting an alias depending on unstable alias for install
For the sanity check calculations, all component aliases are manually
selected to detect possible errors in packaging preventing installing
of an alias. As the order of install for selected aliases is undefined,
in some cases the installer did not detect that an alias requires an
unstable alias.

Consider an example scenario:
* aliasA -> requires aliasB
* aliasB -> requires aliasC
* aliasC -> missing from installer

The unstable status is set the first time we query for the required
aliases of an alias, so the case for "ReferenceToUnstable" could go
amiss if the required aliases list for "aliasA" was populated before
for "aliasB".

Fix by populating the required aliases and components in the order
of the alias dependency graph.

Change-Id: Id2fc9e3e1b315f5be98ca8915a0938abad1a2eb5
Reviewed-by: Katja Marttila <katja.marttila@qt.io>