Enable component script postload setting per repository
This is a custom requirement for Qt Online Installer. Qt Online
Installer has a huge amount of scripts to be evaluated. Most of the
scripts don't need an evaluation unless those are installed. Skipping
the evaluation of scripts has quite significant
performance impact. As the postLoad setting in component level is a big
task, enabling support for Qt Online Installer to set the postLoad value
to scripts per repository.

Task-number: QTIFW-3173
Change-Id: If37d7ff39ad07ee56e9a9559c837966c23d76d6e
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>