ComponentAlias: fix missing data when multiple keys have same value
When populating ComponentAlias data from an alias source file, there
was an oversight with the logic of getting the key by a value when
iterating the AliasData map. For example with the text values, like
DisplayName and Description, there may be same values and the logic
breaks.

Instead use an iterator with the map.

Change-Id: I59c2f45849533d1e664c3e9e2ccd9bba142a7b57
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>