Prevent uninstalling forced component
If forced installation is used, then the item is uncheckable. This was
not correctly saved to component's internal variable. Also although item
is not user checkable, the check state needs to be counted in the model.
Uncheckable items might have for example child items and then the
uncheckable item is also installed (or uninstalled, depending on
selection).

Task-number: QTIFW-2589
Change-Id: I9ea967e34cf3f2048312fb5eacd6d0b397d7215c
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>