Fix EnvironmentVariable undo
If EnvironmentVariable value contains both @TargetDir@ and hardcoded windows
path separators, installer is unable to undo the EnvironmentVariable in
uninstall. Fixed so that the path separators are ignored when comparing
the values in undo.

Another fix would be to use Linux native separators when setting the
value, or resolve @TargetDir@, and calling
installer.toNativeSeparators() before calling addOperation()

Task-number: QTIFW-1148
Change-Id: I2eb8f9a0580e0c3f7e6dbdbef5ecbf6cd8940481
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>