Fix offline generator state not resetting when navigating back
In case the user chose to generate offline installer by clicking the
"Create Offline Installer" button on ComponentSelectionPage, navigating
back to the page did not reset the internal binary marker. This made
it impossible to perform a normal installation afterwards.

Fix by resetting the state on ComponentSelectionPage::entering().

Task-number: QTIFW-3236
Change-Id: I0951ec90e70dff9e8e233dee0cc5592c5c509c28
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>
Reviewed-by: Katja Marttila <katja.marttila@qt.io>