Disable 'clear cache' button until new cache settings are applied
Check the cache validity and use it as the initial state for the button.
After clearing the cache, regardless of success, the cache will become
invalid so the button should stay disabled until new settings are
applied and the cache is reinitialized.

Also disable the button in case the selected cache path does not match
the one currently applied in the settings.

Task-number: QTIFW-3252
Change-Id: I1b403fbc8a6680dfd5222296479ab0821cdb4a37
Reviewed-by: Katja Marttila <katja.marttila@qt.io>