Speed up cache validation
Cache validation takes some time especially in Windows, mostly the time
is spent in opening the files for the integrity check. Instead of
writing the sha checksum to file to verify the integrity, name the file
itself with the sha checksum. Then it is enough to check if the correct
file exists and we don't need to open the file.

Task-number: QTIFW-3328
Change-Id: I36835d783ccf2f6aad8c83e14ec19cbdfc6db71e
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>