Fix maintenance tool register path parameters
In Windows, maintenance tool is written to registry. 'Add or remove'
program in Windows settings uses the registry, and maintenancetool can
be started using the settings. Fixing the maintenancetool parameters so
that selecting 'Modify' will open the maintenancetool in packagemanager
mode, and 'Uninstall' will open the maintenancetool in uninstaller mode.

Change-Id: Ib1e8c80132f6e2eb2f8cfa22910b4ad1ca696f06
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>