Save used CLI arguments for later usage
Qt Installer will send information to backend if unstable components are
found (e.g. dependency is missing), to better track for problems the CLI
arguments are needed as well. Options which are not defined in installer
are not saved, also option arguments which may contain confident
information are not saved either.

Task-number: QTIFW-3070
Change-Id: Iff5bfad725d1f56b7fb6c501138427ef4c58b5c3
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>