3rdparty: update libarchive sources to 3.7.4
Upstream details:
https://github.com/libarchive/libarchive/releases/tag/v3.7.4

Also update related test data archives - there has been a change
in libarchive to translate Windows permissions to Unix permissions
for 7zip archives. The permission mode for normal files is 0644.

The 7zip and QBSP test data archives were originally generated
on Windows, so the test runs on Linux and macOS are affected by the
libarchive change. Update the expected permission mode in tests to
0644 and re-package the other archive formats with same permissions
mode as well.

Change-Id: I791151bc2342655a1bec96bc0f4aafd468519065
Reviewed-by: Katja Marttila <katja.marttila@qt.io>
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>