CLI: Fix fetching components in alias from fallback repositories
Alias components can refer to a individual components which can exist in
fallback repository. If alias is marked as unstable, try fetching the
fallback repositories, if those are not fetched already.

Also resolve only the alias packages that are selected for install.
There is no point in resolving aliases which we are not interested in,
and it also provided unnecessary warning messages if the components
belonging to certain aliases exists in fallback repositories.

Fixed also bug where alias with optional broken dependency was not
installed.

Change-Id: I69a197e652ba83300551e054b41e5932311dfd61
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>