Optimize alias search
Aliases do not need metadata, so we can skip the metadata fetch when
searching for alias packages, which will significantly improve the alias
search.

Also changed the search output so that all aliases are listed despite of
the unstable state, this mimics the behavior we already have
when listing the individual components. That means we can also skip the
sanity check for aliases, which also improves the performance.

Task-number: QTIFW-3392
Change-Id: I129815e1c428eb420f0506e8b141b44fa51f8c4a
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>