Fix macos app bundle update
MacOs app bundle cannot be replaced while it is running, if the app
is used after replace, e.g. operations are called from another component
to be installed.
Fixed so that maintenance tool app bundle can be extracted to a folder in
install script, and function deferredRename, which is called as last step
when running installer, will move the app to root and removes the possible
subfolder which was created during extract operation.

Task-number: QTIFW-3408
Change-Id: I62d873abd3155be533fca2f80afcb820042536d2
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>