binarycreator: increase max file descriptor limit for Linux
This was already done on macOS - increase the maximum open file
descriptor limit on Linux as well. Set the process soft limit
to match the value of the hard limit. This makes errors related to
too many open files more unlikely when assembling the installer
binary's resource layout with huge number of archives.

Task-number: QTIFW-3505
Change-Id: Ie80493b32224672eb720a0bace1b740bd097f3aa
Reviewed-by: Katja Marttila <katja.marttila@qt.io>