CI: Stop making macos IFW tools offline installers
CI has restricted access to private repositories so we need to stop
using those. This will affect the signing an notarization of IFW tools
installer which we deliver to customers. As we cannot sign/notarize the
installer cannot be launched, thus we are not making the installers in
CI systems any more.
This will have no effect users own signing/notarisizing process.

Task-number: QTIFW-3581
Change-Id: Iad922ff90a4e385fce7745b105841ff7a967711b
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>