Move categories selection next to 'Search' bar
Move categories selection to dropdown menu based on UI specs.
Added a new CheckableCombox class so that when selecting a text item in
the combobox, also the checkbox is checked. Also adjusting the size of
the combobox popup, so that view is resized to the content while the
combobox itself stays fixed size.

Task-number: QTIFW-3545
Change-Id: Iac2befb732271952c78439684374f17f7a5fd6da
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>