Add possibility to add information about category usage
Added possibility to productkeycheck extension to add information about
certain category usage. Information is a dialog and must be confirmed
from UI. It is the responsibility of the developer who extends the
productkeycheck to decide if information is needed and the conditions
when it must be shown. Default implementation does nothing.

Task-number: QTIFW-3597
Change-Id: Ieb4dfef33417bc44d6380e9843b5cc778dcb3f4f
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>