Fix  ReadyForInstallation -page layout
If installer is started in verbose mode installer will show detailed
text in the UI. If an extra UI element is added to the page, installer
did not layout the items correctly, resulting the detailed text not to
occupy the space it had available. Fixed so that the detailed text will
be stretched to the available space. Also removed excess layout item
generations.

Task-number: QTIFW-3608
Change-Id: I9c585723e415fb5ae82d7bab61d1560e074a6905
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>