Do not show 'RunProgram' when creating offline installer
'RunProgram', added from configuration file, is shown in the Finished
page as a checkbox. Clicking it will start the program defined in the
configuration file. Usually the program refers to a newly installed
program. When creating offline installer, the program is not installed
but bundled to the offline installer, so the 'RunProgram' should not be
shown when creating offline installer as there is no program to run.

Change-Id: I084656c3399d4f1902f08462fe4a18b423dd43ce
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>