Fix page list text for Start menu -page
The page list text for start menu -page was not defined, which means
that the title text is used. The title text in many langages is longer
than what is fit to the page list -area. This commit creates a shorter
text for page list widget, so the texts won't clip.

Task-number: QTIFW-3646
Change-Id: I012e4a9e6fc122d11470781538a08334653ec0e6
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>