Fix CheckableCombobox layout in RTL
CheckableComboBox draws a transparent pixmap so that the text does not
overlap with the checkbox (BUG PYSIDE-2906). In RTL the pixmap is not
drawn to the correct location making the text to overlap with the
checkbox again. Fixed by layouting the text to the left side of the
widget when RTL is in use.

Task-number: QTIFW-3648
Change-Id: I1693837d287c93643c50b7cb90b76539fed946b9
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>