Component: optimize retrieving install state
Perform an integer comparison to determine the install state of a
component. This is much faster than the previous string comparison,
and avoids constructing temporary QString objects when invoked.

A callgrind analysis revealed that the method got invoked over three
million times when calculating components to install with the Qt Online
Installer, when the user changes their default install selections, in
case all repository categories were enabled. The relative cost of this
for the whole cycle was around 36%, which decreased to so negligible
that it was skipped in the measurement results.

Task-number: QTIFW-3650
Change-Id: I05068b4cd6cbc7a77b1f5f88ac21776516fe0260
Reviewed-by: Katja Marttila <katja.marttila@qt.io>