Remove "coin" subdirectory from exported source archives
The "coin" subdirectory is not needed in the source archives, that
are shipped as part of the official IFW releases.

Also move the create_ifw_installer.py script from "coin" to "dist"
as it is executed as a post-step for the IFW build and needs to be
present on the build VM. The script can be used outside of Coin as
well.

Change-Id: Ib475c3eded481c4757ec8e8571251df35c288dee
Reviewed-by: Katja Marttila <katja.marttila@qt.io>