Fix silent errors on automatic tree name construction
There were two silent errors with an example setup with component
names using following name pattern:

* componentRoot
* component (TreeName=componentRoot.component, moveChildren=true)
* component.subcomponent

"componentRoot" would be incorrectly set with an automatic treename
"componentRoot.component", because it was deemed a child of "component"
due to string comparison checking if its name starts with
"component", which is a component with explicit TreeName. Fix the
comparison to take into account the "." delimiter, which ensures
we only match child components of the component to compare to.

Another issue was that the automatic treename substitution replaced
all occurrences of the substring matching the original name of the
component, so in this case "component.subcomponent" would have a
treename "componentRoot.component.subcomponentRoot.component". Fix
replacing only the first occurrence of the original name at the
start of the component name.

Add auto test case to trigger test failure without the fixes.

Task-number: QTIFW-3670
Change-Id: I01edc5d9657626e321abeb9d1f7e323c218b0a7c
Reviewed-by: Katja Marttila <katja.marttila@qt.io>